PDF ADRIEND DONE + HASH + FIX FRAIS HF
