+++ includes/fct.inc.php
@@ -1,4 +1,5 @@
-        <?php
+<?php
+
 /**
  * Fonctions pour l'application GSB
  *
@@ -19,8 +20,7 @@
  *
  * @return vrai ou faux
  */
-function estConnecte()
-{
+function estConnecte() {
     return isset($_SESSION['idVisiteur']);
 }
 
@@ -29,11 +29,9 @@ function estConnecte()
  *
  * @return vrai ou faux
  */
-function comptableEstConnecte()
-{
+function comptableEstConnecte() {
     return isset($_SESSION['idComptable']);
 }
-
 
 /**
  * Enregistre dans une variable session les infos d'un visiteur
@@ -44,8 +42,7 @@ function comptableEstConnecte()
  *
  * @return null
  */
-function connecter($idVisiteur, $nom, $prenom)
-{
+function connecter($idVisiteur, $nom, $prenom) {
     $_SESSION['idVisiteur'] = $idVisiteur;
     $_SESSION['nom'] = $nom;
     $_SESSION['prenom'] = $prenom;
@@ -60,21 +57,18 @@ function connecter($idVisiteur, $nom, $prenom)
  *
  * @return null
  */
-function comptableConnecter($idComptable, $nom, $prenom)
-{
+function comptableConnecter($idComptable, $nom, $prenom) {
     $_SESSION['idComptable'] = $idComptable;
     $_SESSION['nom'] = $nom;
     $_SESSION['prenom'] = $prenom;
 }
-
 
 /**
  * Détruit la session active
  *
  * @return null
  */
-function deconnecter()
-{
+function deconnecter() {
     session_destroy();
 }
 
@@ -86,8 +80,7 @@ function deconnecter()
  *
  * @return Date au format anglais aaaa-mm-jj
  */
-function dateFrancaisVersAnglais($maDate)
-{
+function dateFrancaisVersAnglais($maDate) {
     @list($jour, $mois, $annee) = explode('/', $maDate);
     return date('Y-m-d', mktime(0, 0, 0, $mois, $jour, $annee));
 }
@@ -100,8 +93,7 @@ function dateFrancaisVersAnglais($maDate)
  *
  * @return Date au format format français jj/mm/aaaa
  */
-function dateAnglaisVersFrancais($maDate)
-{
+function dateAnglaisVersFrancais($maDate) {
     @list($annee, $mois, $jour) = explode('-', $maDate);
     $date = $jour . '/' . $mois . '/' . $annee;
     return $date;
@@ -114,8 +106,7 @@ function dateAnglaisVersFrancais($maDate)
  *
  * @return String Mois au format aaaamm
  */
-function getMois($date)
-{
+function getMois($date) {
     @list($jour, $mois, $annee) = explode('/', $date);
     unset($jour);
     if (strlen($mois) == 1) {
@@ -133,8 +124,7 @@ function getMois($date)
  *
  * @return Boolean vrai ou faux
  */
-function estEntierPositif($valeur)
-{
+function estEntierPositif($valeur) {
     return preg_match('/[^0-9]/', $valeur) == 0;
 }
 
@@ -145,8 +135,7 @@ function estEntierPositif($valeur)
  *
  * @return Boolean vrai ou faux
  */
-function estTableauEntiers($tabEntiers)
-{
+function estTableauEntiers($tabEntiers) {
     $boolReturn = true;
     foreach ($tabEntiers as $unEntier) {
         if (!estEntierPositif($unEntier)) {
@@ -163,8 +152,7 @@ function estTableauEntiers($tabEntiers)
  *
  * @return Boolean vrai ou faux
  */
-function estDateDepassee($dateTestee)
-{
+function estDateDepassee($dateTestee) {
     $dateActuelle = date('d/m/Y');
     @list($jour, $mois, $annee) = explode('/', $dateActuelle);
     $annee--;
@@ -180,8 +168,7 @@ function estDateDepassee($dateTestee)
  *
  * @return Boolean vrai ou faux
  */
-function estDateValide($date)
-{
+function estDateValide($date) {
     $tabDate = explode('/', $date);
     $dateOK = true;
     if (count($tabDate) != 3) {
@@ -205,8 +192,7 @@ function estDateValide($date)
  *
  * @return Boolean vrai ou faux
  */
-function lesQteFraisValides($lesFrais)
-{
+function lesQteFraisValides($lesFrais) {
     return estTableauEntiers($lesFrais);
 }
 
@@ -222,8 +208,7 @@ function lesQteFraisValides($lesFrais)
  *
  * @return null
  */
-function valideInfosFrais($dateFrais, $libelle, $montant)
-{
+function valideInfosFrais($dateFrais, $libelle, $montant) {
     if ($dateFrais == '') {
         ajouterErreur('Le champ date ne doit pas être vide');
     } else {
@@ -232,7 +217,7 @@ function valideInfosFrais($dateFrais, $libelle, $montant)
         } else {
             if (estDateDepassee($dateFrais)) {
                 ajouterErreur(
-                    "date d'enregistrement du frais dépassé, plus de 1 an"
+                        "date d'enregistrement du frais dépassé, plus de 1 an"
                 );
             }
         }
@@ -254,8 +239,7 @@ function valideInfosFrais($dateFrais, $libelle, $montant)
  *
  * @return null
  */
-function ajouterErreur($msg)
-{
+function ajouterErreur($msg) {
     if (!isset($_REQUEST['erreurs'])) {
         $_REQUEST['erreurs'] = array();
     }
@@ -267,8 +251,7 @@ function ajouterErreur($msg)
  *
  * @return Integer le nombre d'erreurs
  */
-function nbErreurs()
-{
+function nbErreurs() {
     if (!isset($_REQUEST['erreurs'])) {
         return 0;
     } else {
