+++ nbproject/project.properties
@@ -1,5 +1,5 @@
 include.path=${php.global.include.path}
-php.version=PHP_73
+php.version=PHP_74
 source.encoding=UTF-8
 src.dir=.
 tags.asp=false
