+++ script/hashpwd.php
@@ -1,0 +1,10 @@
+<?php
+
+/* 
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+require_once '../includes/class.pdogsb.inc.php';
+PdoGsb::hashPasswordsVisiteurs();
+PdoGsb::hashPasswordsComptables();
