+++ styles/style.css
@@ -53,4 +53,4 @@
 .pdfstyle{
     width: 35px;
     height: 50px; display : block; margin : 10px;
-}
+}
